- Style fixes (use lower camel case in all cases)
- Added TYPE_prefix for event.type strings
- Removed commented code
- Removed "database.selectLastEvents" because it is not cross-database. Do the same via code.

+++ src/org/traccar/api/resource/EventResource.java
@@ -32,7 +32,7 @@ public class EventResource extends BaseResource {
     @GET
     public Collection<Event> get(
             @QueryParam("deviceId") long deviceId, @QueryParam("type") String type,
-            @QueryParam("interval") long interval) throws SQLException {
+            @QueryParam("interval") int interval) throws SQLException {
         Context.getPermissionsManager().checkDevice(getUserId(), deviceId);
         return Context.getDataManager().getLastEvents(deviceId, type, interval);
         }
